Rename beta confidence bands for the CDF to LD confidence bands
The beta confidence bands were first introduced in the paper:
"A Probabilistic Upper Bound on Differential Entropy" by
Learned-Miller and DeStefano (2008). The beta distribution is
common and many statistics could relate to it. In contrast, naming
the confidence bands after the creators is much more specific and
less likely to lead to confusion.