Make the correct coverage unit tests faster and more robust
Reduce the number of samples and trials in the unit tests that
test that the confidence bands have the correct coverage. In
addition, look for violations just before and after the jump
discontinuities which both makes the tests faster and makes them
more robust in that they find more of the confidence band
violations. Finally, use a Clopper-Pearson confidence interval in
place of a normal approximation for setting the bounds against
which we test the empirically measured coverage. The normal
approximation interval has poor coverage for high levels of
confidence (e.g., 99% confidence).