Match parameters for test_confidence_bands to other unit tests
Change the number of samples and reduce the number of confidence
levels that are tried in test_confidence_bands in order to speed
it up. In addition, match those parameters with ones from the
other confidence band test cases since this will then speed up the
tests as the values for the confidence band weights are cached.