Clean up the code for beta_highest_density_interval
Give names to intermediate quantities in
beta_highest_density_interval to make the code more readable. Some
of the transformations also speed up the code a bit, but the
difference is relatively negligible because computing the upper
endpoint, y, using the special functions for the beta density's
cdf and ppf is very expensive and dominates the runtime.