Rename the project from ERSA to OPDA
The project has expanded beyond its original scope of just
analyzing and extrapolating random search. The new scope of the
project will be to create a suite of analyses for determining and
understanding optimal designs (of deep learning models). As a
result, rename the project from ERSA (extrapolated random search
analysis) to OPDA (optimal design analysis).