Improve the exit condition for beta_highest_density_interval
beta_highest_density_interval uses binary search to find highest
probability density intervals for beta distributions. Binary search
cuts the bracket size in half each iteration, so it makes a known
improvement to the current answer. As such, we can precompute the
number of iterations necessary to achieve a given tolerance.

Update the implementation of beta_highest_density_interval to
precompute the number of iterations instead of checking when a
tolerance has been reached. This change both avoids the need to
ever raise an exception due to lack of convergence and speeds up
the code because we no longer need to check the exit condition
each iteration.