Make EmpiricalDistribution.v_tuning_curve handle floats properly
The v statistic tuning curve estimate is only defined for (positive)
integers. Thus, if v_tuning_curve is called on a float, the
function should convert the float to an int if the float is a
natural number; otherwise, the function should raise a ValueError.
Update the documentation, implementation, and test cases to match
this behavior.