Move setup.py and requirements files into pyproject.toml
Move all of the information from setup.py and the requirements
files (requirements.txt and dev-requirements.txt) into
pyproject.toml so that we only need one file for configuring the
packaging. In addition, update some of the packaging information
either because the options were obsolete or the newer information
better describes the project.