Fix the plotting example in README.rst
Fix the plotting example in README.rst so that it corresponds to
what the output would be in a shell session. The default
(interactive) backend can cause issues in a shell session, so use
the (static) "AGG" backend instead. With these changes, the example
now also passes when executed as a doctest.