Fix flakiness in EmpiricalDistributionTestCase.test_quantile_tuning_curve
The test is flaky because the quantile tuning curve estimated via
sampling can easily be off by one order statistic. In the checks
for when EmpiricalDistribution is given a sample with duplicates,
the difference between order statistics is 50 rather than 25,
while the check asserts that the difference should be at most 25.
Fix the issue by changing the input to reduce the difference to 25.