Fix bug in EmpiricalDistribution.confidence_bands
The bug in EmpiricalDistribution.confidence_bands causes the bands
to be conservative, especially so for small sample sizes (less than
5 or so). Even with the bug, the bands' coverage is close to exact
for large samples.

The bug has two components. First, the hidden helper functions
(e.g., _dkw_band_weights) should return lists of n + 1 weights when
given n samples. These helper functions need to return an extra
weight because otherwise it is unclear how much probability mass is
on the lower bound of the distribution's support versus the
smallest order statistic. For example, consider the DKW band with
one sample. The upper confidence band is epsilon before the sample
and 1 after the sample; however, if we only provide the cumulative
weight (CDF value) *at* the sample, then we have no way of knowing
what epsilon was. So, it is necessary to report the CDF value
*before* the first sample and thus have one extra weight.

The second part of the bug is that the cumulative weights returned
by the helper functions were not correctly transformed into weights
for each sample. The correct way to do this is to prepend 0 and
append 1 on the list and then difference it. This process will
assign the correct amount of probability mass to the lower and
upper bounds on the support as well as the samples.

In addition, improve the coverage tests so that they catch the
aforementioned bug. First, remove the coverage tolerances since
they are actually unnecessary once the bug has been fixed. Second,
the bug is most apparent on small samples, so test coverage with a
sample size of 2 instead of 5. The sample size must be at least 2
in order for the LD highest density bands to be well-defined (since
the highest density interval for the beta distribution is only
well-defined in this case).