Sample weights better in EmpiricalDistributionTestCase.test_quantile_tuning_curve
When the weights are sampled uniformly from the unit simplex, the
weight for a particular sample can be very small. This small weight
makes it difficult to approximate the tuning curve accurately
enough to ensure that the test passes (or that the estimated tuning
curve's value at each point is either the correct order statistic
or its neighbor). To reduce the test's flakiness, concentrate the
dirichlet distribution towards the center a bit when sampling the
weights in order to avoid the extremes of the simplex.

In addition, speed up the test by reducing the sample size used to
estimate the quantile tuning curve. After avoiding the extremely
small weights, the sample size can be reduced without making the
test too flaky.