Improve QuadraticDistribution.cdf on inputs outside of the support
When ys contains values outside the distribution's support, numpy
logs warnings for taking powers on invalid inputs. Clipping the
inputs to be inside of the support produces correct output values
because values below the support will map to a where the CDF is 0
and, similarly, values above the support will map to b where the
CDF is 1. Moreover, clipping the inputs avoids the warning and
requires less computation.