Improve argument pre-processing in QuadraticDistribution.ppf
The input, qs, to QuadraticDistribution.ppf must be between 0 and
1. The ppf method allows inputs to be slightly less than 0 or
greater than 1 because such small deviations may be due to rounding
errors. This flexibility means that qs should be clipped to 0 and
1 at the start of the function. Currently, the output is
post-processed to the same effect (so there is no bug in the code),
but that's less clear and efficient than clipping qs directly.