Migrate tests off numpy's legacy API for random numbers
Global RandomState is now considered a legacy API in numpy. The
new Generator API is faster and will continue to receive updates.
In addition, using generators instantiated with random seeds
rather than a global state creates the opportunity to log these
random seeds on test failures so that the failure can be
reproduced. Thus, create a new base class for random test cases,
RandomTestCase, that generates a random seed, logs its value, and
instantiates a generator for each test. Then, migrate appropriate
tests to use this base class. The random seed is logged at the INFO
level, so set pytest's logging level to INFO by default so that the
random seed will be printed in the event of a test failure.