Disallow None for the a and b arguments to EmpiricalDistribution
Allowing None as a value for a and b creates multiple
representations for the same thing. Namely, an upper bound of
infinity can be represented by passing either np.inf or None for b.
Similarly, a lower bound of negative infinity can be represented by
passing either -np.inf or None for a. Having one representation
leads to simpler code and will make it easier to check equality
between EmpiricalDistribution instances.