Define package-data for setuptools in pyproject.toml
The default matplotlib style for experiments is not included in the
experiments package. Due to this omission, the notebooks can not be
run from a regular install of the packages because the notebooks
require this default style. Only editable installs work because
editable installs do include the default style file.

Define the package-data attribute for setuptools in pyproject.toml
in order to include the default.mplstyle file and fix this issue.