Prune the optional dependencies in pyproject.toml
The optional dependencies are frequently installed in nox sessions
and continuous integration. Thus, it is beneficial to have each set
of optional dependencies be minimal for the desired use case. Only
the notebooks (nbs) and testing (tests) optional dependencies
require the experiments package. It can be removed from the rest.
After removing the experiments package from the documentation
(docs) dependencies, matplotlib must be added back in, as it is no
longer included transitively.