Split local and distribution package builds via local-only comments
Some code, like the experiments package, is only intended for use
in the research behind opda. Such code is valuable to keep in this
repository so that the research can be reproduced; however, it is
not useful to distribute that code to end-users who only want to
use the methods. Thus, split the local and distribution package
builds for these two different use cases. Implement this split
using comments in the pyproject.toml file. Each line ending in a
`# build: local-only` comment should be removed before building the
distribution package. Similarly, everything after the header:
`# -- build: local-only -----` should be removed.

Using these comments, the distribution package's pyproject.toml
file can be constructed with simple text processing. For example,
using awk:

  awk \
    '/# -- build: local-only --/ {exit} !(/# build: local-only/) {print}' \
    pyproject.toml

Even using this mechanism to exclude the experiments package,
setuptools will still include the source for it in the sdist. So,
also create a MANIFEST.in file intended solely for the distribution
package and adapt pyproject.toml to make sure it builds the local
and distribution packages correctly (excluding all local-only files
from the distribution package's sdist and wheel).