Disable lint rules: ERA (Eradicate)
Disable the ERA lint rules because they give too many false
positives. The ERA rules detect commented out code that shouldn't
get committed. Unfortunately, they also fire when code is used as
an example or to concisely express something like an equation in a
regular comment.

The rules are already disabled for tests and notebooks. Remove the
rules completely to prevent false positives in the main code.