Fix QuadraticDistribution for the case when a equals b
When a equals b, the quadratic distribution is a point mass. Its
PDF is zero everywhere except at the point a == b, where it's
infinite. The CDF is zero below a == b and 1 at and above it. The
PPF always returns a == b, and so on.

Currently, this case is untested and the QuadraticDistribution
class does not handle it correctly. Add tests and update the class
for this case.