Add NoisyQuadraticDistribution._chebyshev_coefficients
Add the private method _chebyshev_coefficients to the
NoisyQuadraticDistribution. This method computes the coefficients
of the Chebyshev approximation to x^k on a given interval.

It's necessary to compute a partial fractional normal moment for
x^-0.5 when computing the PDF of the noisy quadratic distribution
with c = 1. There is no minimax approximation to x^-0.5 on 0 to 1
because it goes to infinity at 0. In many cases, this moment can
be computed by using a recursive formula to step down from the
higher moments (0.5 and 1.5); however, when the scale (signed
standard deviation) of the normal distribution is small, stepping
down cancels out most of the significant digits from the higher
moments, leaving just the approximation error.

Luckily, when the scale is small, we can compute the partial
fractional normal moment using an idea similar to the delta method
in statistics. When the normal distribution's standard deviation is
small, most of the probability mass is close to the mean. Thus, we
can approximate the expectation of a function, E[f(X)], by
replacing f in the integral with an approximation that's good for
several standard deviations, o, about the mean, u. For example,
u - 6o to u + 6o. The delta method uses a Taylor approximation at
u. Empirically, Taylor approximations don't work well for our case
because they fail to capture the asymmetry of x^-0.5 due to the
fact that x^-0.5 goes to infinity at 0. Instead we use a Chebyshev
approximation over the whole interval, which provides a much better
fit.

Thus, when the scale is small, we need to compute the coefficients
of Chebyshev approximations on the fly in order to compute the
partial fractional normal moment for x^-0.5 and thus the PDF of the
noisy quadratic distribution when c = 1.