Add NoisyQuadraticDistribution._base_partial_normal_moments
In order to compute the noisy quadratic distribution's PDF and CDF,
the NoisyQuadraticDistribution class will compute the partial
normal moments using a recursive formula. The best base case for
this formula depends on the distribution's parameters, thus the
logic for selecting the best base case is somewhat complex. So, use
a dedicated method to determine the base case.

The trickiest case is the partial moment for k = -0.5. Either you
can step down to this moment from higher ones (e.g., 0.5 and 1.5)
using the recursive formula, or you can compute it directly using a
Chebyshev approximation. No minimax polynomial exists for k = -0.5
because the function goes to infinity at 0. Stepping down from the
higher moments involves a subtraction that can remove most or all
of the significant digits when the scale is small. Thus, step down
to this moment when the scale is sufficiently large, and compute it
directly when the scale is not. For more information, see commit
a99f06c (Add NoisyQuadraticDistribution._chebyshev_coefficients,
2024-02-11).