Add NoisyQuadraticDistribution.pdf
Compute the PDF for the noisy quadratic distribution using the
the partial fractional moments of the normal distribution. Use this
approach instead of direct numerical integration of the convolution
of the quadratic and normal distribution's densities because
numerical integration is slower and has too much error. Since the
quadratic distribution's density can go to infinity at the boundary
when c = 1, numerical integration fails to accurately compute the
density there.