Add the NoisyQuadraticDistribution._approximate_with attribute
We will compute the CDF and PDF of the noisy quadratic distribution
using partial fractional moments of the normal distribution. This
strategy works well for most parameter values; however, it suffers
from numerical issues when o / (b - a) is very small or very large.
In these cases, use limiting approximations to the distribution.
When o / (b - a) is small, the distribution is approximately the
(noiseless) quadratic distribution. When o / (b - a) is large, the
distribution is approximately normal. When a = b and o = 0, the
distribution is a point mass at a = b.