Add NoisyQuadraticDistribution.cdf
Compute the CDF for the noisy quadratic distribution using the
the partial fractional moments of the normal distribution. Use this
approach instead of direct numerical integration of the convolution
of the quadratic distribution's CDF and the normal distribution's
PDF because numerical integration is much slower. For small inputs
(e.g., when ys is a scalar), either numerical integration or the
normal moments approach can be faster depending on the input and
the distribution's parameters. In some quick benchmarks, the normal
moments approach was about 3 times slower to 6-7 times faster.
However, the normal moments approach scales much better. For large
inputs (e.g., arrays of 10,000), the normal moments approach was
about 100 to 725 times faster than numerical integration using the
composite trapezoid rule (and for a similar level of accuracy).
This scalability is important since the average tuning curve will
be computed via numerical integration of the CDF (and thus will
evaluate the CDF at many different points).