Throw error if fraction too small in QuadraticDistribution.estimate_initial_parameters_and_bounds
Throw an error if the fraction is small enough that the sample used
to estimate the initial parameters and bounds is empty. Throwing an
error in this case also fixes a bug. Currently, if convex = False
and fraction is small enough that int(fraction * len(ys)) = 0, then
all of ys is used to estimate the initial parameters and bounds
since np.sort(ys)[-0:] == np.sort(ys). By throwing an error, we
avoid this edge case and fix the bug.