Make scale always positive in opda.parametric.NoisyQuadraticDistribution
The formulas for the noisy quadratic distribution involve a "scale"
parameter. Currently, this scale parameter can be positive or
negative and represents a multiplier on a standard normal
distribution.

Allowing the scale parameter to be negative makes the code more
difficult to understand for two reasons. First, negative scale
parameters for the normal distribution are unfamiliar to most
people and thus more difficult to understand. Second, the
derivation involving the scale parameter requires you to consider
in which direction you're integrating the real line when you take
expectations. This is a pretty subtle issue.

In order to make the code easier to understand, require that the
scale parameter is always positive (i.e., the standard deviation)
and adjust the formulas accordingly.