Skip tests when they fail due to an upstream bug in scipy 1.14.0
On Linux, scipy 1.14.0 introduced a bug where stats.beta.ppf
incorrectly returns NaN for certain arguments:

  >>> from scipy import stats
  >>> stats.beta(5., 5.).ppf(0.5)
  np.float64(nan)

This bug does not occur for most argument values and thus is
somewhat rare to encounter. For more information, see
https://github.com/scipy/scipy/issues/21303.

Still, the bug causes some of opda's tests to fail. Since the bug
is in scipy rather than opda, skip the failing tests when the bug
is present.