Correct the quantile function's definition in ppf docstrings
The ppf method of NoisyQuadraticDistribution computes the quantile
function by taking the infimum over the distribution's support. In
particular, when o = 0 the support ranges from a to b and
evaluating ppf at 0 gives a (not negative infinity). The docstring
for ppf incorrectly states that the infimum quantifies over the
real line (which is only true when o > 0).

Correct the quantile function's definition by specifying that the
infimum quantifies over the support. In addition, update the
docstring for QuadraticDistribution to also quantify over the
support in order to make it consistent with
NoisyQuadraticDistribution.