Use distribution approximations in NoisyQuadraticDistribution.ppf
The ppf method of NoisyQuadraticDistribution is implemented via
numerical inversion of the CDF. This numerical inversion is
unnecessary for certain parameter values, for example when the
point mass, noiseless, and normal approximations apply.

In addition, when c = 1 and o = 0, small differences in y create
large changes in F(y) near either a or b because the PDF goes to
infinity at one of them. Thus, while the numerical inversion gives
high accuracy in terms of |y - F^-1(q)|, it gives a lower accuracy
in terms of |F(y) - q|. This lower accuracy can cause the test:
NoisyQuadraticDistributionTestCase.test_ppf_is_inverse_of_cdf to
fail. For example, consider the following code:

  >>> from opda.parametric import NoisyQuadraticDistribution
  >>> dist = NoisyQuadraticDistribution(0., 1., 1, 0., True)
  >>> abs(dist.cdf(dist.ppf(1e-7)) - 1e-7) < 1e-5
  False

Using the noiseless approximation increases the accuracy in this
case, preventing the issue.