Make NoisyQuadraticDistributionTestCase.test_ppf_is_inverse_of_cdf more robust
Sometimes NoisyQuadraticDistributionTestCase.test_ppf_is_inverse_of_cdf
will fail because it samples a value of y such that the cdf method
returns 1.0. When this happens and o > 0, the ppf method (correctly)
maps that point to infinity. Thus, the difference from y is also
infinite, causing the test to fail.

To avoid this case, check that each sampled value maps to a CDF
value strictly between 0 and 1 before testing it.