Fix bug in quantile_tuning_curve method of parametric distributions
Fix the quantile_tuning_curve methods of QuadraticDistribution and
NoisyQuadraticDistribution. Both methods use an incorrect formula
for the quantile tuning curve when minimize is True. The tests do
not catch this bug because they only test q = 0.5 and the incorrect
formula happens to work for this one case. In addition, the
test_quantile_tuning_curve tests actually check non-integral
arguments to the quantile tuning curve using an invariant which
would fail when this bug is present, except the invariant was also
entered into the test incorrectly.

To resolve these issues:

  * Add a regression test to QuadraticDistributionTestCase and
    NoisyQuadraticDistributionTestCase to catch similar bugs in the
    future: test_quantile_tuning_curve_with_different_quantiles.
  * Fix the invariant used to test non-integral arguments in
    test_quantile_tuning_curve on QuadraticDistributionTestCase and
    NoisyQuadraticDistributionTestCase so that the tests will pass
    when the quantile_tuning_curve methods are correct.
  * Fix the bug in QuadraticDistribution.quantile_tuning_curve and
    NoisyQuadraticDistribution.quantile_tuning_curve.

This bug does not affect EmpiricalDistribution.quantile_tuning_curve
or its tests as they all use the correct formula.