Add C_MIN and C_MAX class attributes to QuadraticDistribution
Add C_MIN and C_MAX class attributes to QuadraticDistribution and
use them to restrict the range of supported values for c. While the
formulae used to implement QuadraticDistribution theoretically work
for any value of c, we only test a limited range. In addition, when
we implement a fit method for QuadraticDistribution, it would
require extra logic to handle an unbounded range for c. Since the
cases of greatest interest are when c is small (as otherwise random
search fails as an optimization method), supporting larger values
is not worth the added complexity.

Thus, add C_MIN and C_MAX attributes to QuadraticDistribution and
restrict the range of supported values for c to 1 to 10, the same
range as NoisyQuadraticDistribution.