Add QuadraticDistribution.fit
Add a class method for fitting a quadratic distribution to data and
returning it as an instance. The fit method must:

  1. *Fit* a quadratic distribution to data, and optionally
  2. *Censor* the tails of that data, and
  3. *Constrain* the fitted parameters.

For the deeper rationale behind fit's interface and implementation, see
commit 2f6ddc6 (Add NoisyQuadraticDistribution.fit, 2024-09-04).

A few implementation details differ between QuadraticDistribution.fit
and NoisyQuadraticDistribution.fit. Most notably: for the
QuadraticDistribution, the parameters a and b must respectively be less
than or equal to and greater than or equal to all the observations.
Thus, a is bounded from above by the sample minimum and b from below by
the sample maximum. As a result, a and b have disjoint search spaces so
that we always have a < b. So, there's no need to specifically treat the
case where this condition is not satisfied. In addition, the initial
populations differ a bit due to the quadratic distribution only needing
to estimate a, b, c, and convex, not o. Thus, differently shaped initial
populations work better for each method.