Disable lint rules: B023, PLW2901, and SIM102
Disable the following lint rules:

  - B023   : Function definition does not bind loop variable {name}
  - PLW2901: Outer {outer_kind} variable {name} overwritten by inner
             {inner_kind} target
  - SIM102 : Use a single `if` statement instead of nested `if`
             statements

These rules all give false positives, so disable them in order to keep
the linter's noise to a minimum.

B023 seems to fire even when loop variables are not used. Moreover, one
may want to define a function which uses a loop variable but is only
executed in the context of the current loop iteration.

PLW2901 fires even when a loop variable is rewritten using itself, for
example when you loop over a list and want to coerce the loop variable's
type.

SIM102 fires when there are nested if statements that could be replaced
with a single if statement; however, the appropriate call in such
situations could go either way. For example, if using the nested if
statement better matches the code's larger structural pattern then it
can be right choice.