Make NoisyQuadraticDistribution.fit more robust to numerical errors
In practice, NoisyQuadraticDistribution.cdf is not monotonic due to
approximation errors. This non-monotonicity can cause nearby points to
have zero or negative spacings, even at the true parameters.

Currently, fit forces monotonicity by taking a cumulative maximum of the
CDF values. This approach works well unless all of the continuous
parameters (a, b, and o) are fixed, in which case it can fail. The
cumulative maximum sends the loss to infinity whenever there's a
negative spacing. When the best parameters have a negative spacing, the
loss becomes infinite at them, but if at least one continuous parameter
is not fixed then the optimizer can find nearby parameters for which the
spacing is not negative; however, if all continuous parameters are
fixed, then there are no nearby parameters, thus the optimizer will
instead choose the wrong value for c or convex.

To see this issue, set the seed in tests.testcases.RandomTestCase.setUp
to 22174092375427482142500722275170997077. On the previous commit, make
the following change:

       def setUp(self):
           """Configure a random seed and generator for random numbers."""
           # Create the random seed and generator.
  -        self.seed = np.random.SeedSequence().entropy
  +        self.seed = 22174092375427482142500722275170997077
           self.generator = np.random.default_rng(self.seed)

           # Use the random seed and generator globally.

Then, run test_fit_applies_constraints_correctly from
tests.opda.test_parametric.NoisyQuadraticDistributionTestCase. The test
will fail for the case where a, b, c, and o are all fixed. If you
investigate why, you'll see that it's due to two points in the sample
being close enough together so as to have a negative spacing, causing
the optimizer to choose the wrong value for "convex".

To solve the issue, sort the CDF values before computing the spacings
instead of taking the cumulative maximum. When all spacings are
positive, the CDF values will already be in sorted order, thus the
sorting will have no effect. When some spacings are negative, sorting
the CDF values will put the lower values where the correct CDF is lower
and the higher values where it is higher. Thus, monotonicity will always
be satisfied and intuitively the average approximation error should be
reduced.