Adjust the shape parameter grid in NoisyQuadraticDistribution.fit
Occasionally fit fails to optimize the objective when the shape
parameter (o / (b - a)) is around 1e-1. This shape parameter influences
the distribution much more heavily when it's medium sized than when it's
large or small. The intuition is as follows: when o is small relative to
b - a, then the distribution looks like a (noiseless) quadratic
distribution; when o is large relative to b - a, the distribution looks
like a normal; however, when o is medium sized, then it has a large
influence on the distribution's shape.

To improve the fit, adjust the shape parameter's grid to have more
values in the medium sized range.