Fix bug in noxfile.fetch_supported_python_versions
The fetch_supported_python_versions function parses HTML from the
python.org downloads page in order to determine the currently
supported python versions. This function checks a table on that page
for all versions not designated as "prerelease". Currently,
Python 3.8 is designated as "end of life, last release was 3.8.20".
This designation is not "prerelease" but also does not represent
a supported version.

Part of the challenge is that the table is not programmatically
generated but rather intended for human consumption. I am not aware
of a machine readable list of supported python versions, so we have
to parse this less structured data.

To make fetch_supported_python_versions more robust, check both the
release status and its support dates. For release status, consider
only "bugfix" and "security" as supported release statuses. For the
support dates, check that the current date falls within the support
range. If the indicators disagree on the support status, throw an
error as the parsing has probably encountered difficulties.