Improve tests for NoisyQuadraticDistribution.fit's use of random seeds
The test test_fit_defaults_to_global_random_number_generator in
NoisyQuadraticDistributionTestCase failed in continuous integration
by giving the same fitted distribution even when the random seed
was not fixed. This spurious failure occurred because the fits
just happened to be the same even with the randomness.

Re-running test_fit_defaults_to_global_random_number_generator, it
can pass 1,000s of times before encountering a failure. Thus, the
test is not very flaky. The other similar test,
test_fit_is_deterministic_given_generator_argument, also passes
1,000s of times before a failure.

To reduce the occurrence of spurious failures: automatically retry
inequality checks upon the first failure. Retrying on failure
squares the probability of spuriously failing the test; thus, if
the current probability is 1 in 1,000, the new probability would be
1 in 1,000,000. In addition, using the "or" construct will only run
the second try if the first one fails (due to the fact that "or"
short circuits in python). As a result, the retry will only make
the test run longer in the unlikely event that the first try fails.