Reduce flakiness in NoisyQuadraticDistributionTestCase.test_fit
The test_fit test in NoisyQuadraticDistributionTestCase has failed
spuriously several times in continuous integration. The failures
come from the error thresholds for the fitted parameters being too
tight.

To avoid spurious failures, loosen the error thresholds in test_fit
and loosen the corresponding thresholds in other tests of the fit
method for NoisyQuadraticDistribution.