Replace CDF error threshold in NoisyQuadraticDistributionTestCase.test_fit
Replace the threshold for failing on a discrepancy between the true
and the estimated CDFs in test_fit on NoisyQuadraticDistributionTestCase.

The existing threshold uses the DKW inequality to bound how far the
empirical CDF can be from the true CDF and thereby how far the
estimated CDF should be from the true CDF (assuming the difference
between the estimated and the empirical CDF is small). However, the
estimated CDF can be on the opposite side of the empirical CDF as
the true CDF.

To account for this difference, multiply the DKW bound by 2**0.5 as
a heuristic (2**0.5 is the inflation in the standard deviation of a
sum of two independent, identically distributed random variables).
This change also enables us to use 1 - 1e-6 for the confidence
threshold in the DKW bound, better matching the rest of the test
code.