Reduce flakiness in test_fit_applies_limits_correctly tests
Adjust the error thresholds in test_fit_applies_limits_correctly in
QuadraticDistributionTestCase and NoisyQuadraticDistributionTestCase
to reduce the tests' flakiness.

Currently, the tests are not very flaky. For example, the test
test_fit_applies_limits_correctly from QuadraticDistributionTestCase
failed only 1 time in 1,000 runs. Still, the test is not reliable
enough given how often all the tests are run.

The failures are due to expected estimation error for the "a"
parameter when the data is left-censored. The estimation problem
is difficult, so loosen the error threshold for "a" when testing
that the limits interval is left-open. At the same time, the "b"
parameter is actually much easier to estimate in this context, so
tighten the error threshold for "b".