Fix QuadraticDistribution.fit when all ys are negative
When using maximum spacing estimation, the method rounds the data
in order to make the fit more robust. For this rounding, the number
of decimals is chosen based on ys' floating point format. In
particular, the decimals are based on the largest spacing of any of
the values in ys.

When a number is negative, the value returned by np.spacing is also
negative, so the largest value should be taken over the absolute
values of the numbers returned by np.spacing.